fixed styling artifacts
