+++ packages/react-composites/src/composites/common/styles/ParticipantContainer.styles.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { IStackStyles, IStackItemStyles, IStackTokens } from '@fluentui/react';
+import { IStackStyles, IStackItemStyles, IStackTokens, mergeStyles } from '@fluentui/react';
 
 /**
  * @private
@@ -74,3 +74,42 @@ export const peopleSubheadingStyle: IStackItemStyles = {
 export const peoplePaneContainerTokens: IStackTokens = {
   childrenGap: '0.5rem'
 };
+
+/**
+ * @private
+ */
+export const participantListWrapper = mergeStyles({
+  width: '20rem',
+  // max width at 50% of view so the People Pane is not squeezing the Message Pane to almost nothing when on small screen or high zoom in
+  maxWidth: '50vw',
+  height: '100%'
+});
+
+/**
+ * @private
+ */
+export const participantListContainerPadding = { childrenGap: '0.5rem' };
+
+/**
+ * @private
+ */
+export const listHeader = mergeStyles({
+  fontSize: '1rem',
+  margin: '1rem'
+});
+
+/**
+ * @private
+ */
+export const participantListStack = mergeStyles({
+  height: '100%'
+});
+
+/**
+ * @private
+ */
+export const participantListStyle = mergeStyles({
+  height: '100%',
+  overflowY: 'auto',
+  overflowX: 'hidden'
+});
